# openapi.json
{"openapi":"3.0.2","info":{"title":"Lending API","description":"API for managing loans","version":"0.1"},"paths":{"/":{"get":{"summary":"Redirect To Docs","operationId":"redirect_to_docs__get","responses":{"307":{"description":"Successful Response"}}}},"/users":{"get":{"tags":["Users"],"summary":"List Users","operationId":"list_users_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response List Users Users Get","type":"array","items":{"$ref":"#/components/schemas/UserSchema"}}}}}}},"post":{"tags":["Users"],"summary":"Create User","operationId":"create_user_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSchemaBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/loans":{"get":{"tags":["Users"],"summary":"List User Loans","description":"Route to list all loans a user has access to\nThis includes loans both owned by and shared with the user","operationId":"list_user_loans_users__user_id__loans_get","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response List User Loans Users  User Id  Loans Get","type":"array","items":{"$ref":"#/components/schemas/LoanSchema"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/loans":{"post":{"tags":["Loans"],"summary":"Create Loan","description":"Route to create a new loan\nAdds the owner to the loan access map","operationId":"create_loan_loans_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoanSchemaBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoanSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/loans/{loan_id}":{"get":{"tags":["Loans"],"summary":"Get Loan Schedule","description":"Route to return the full monthly loan schedule\nThe loan must be owned by or shared with the user","operationId":"get_loan_schedule_loans__loan_id__get","parameters":[{"required":true,"schema":{"title":"Loan Id","type":"integer"},"name":"loan_id","in":"path"},{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Loan Schedule Loans  Loan Id  Get","type":"array","items":{"$ref":"#/components/schemas/LoanScheduleSchema"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Loans"],"summary":"Update Loan","description":"Updates the loan at the provided loan_id with the new loan_data\nThe loan must be owned by the user at user_id","operationId":"update_loan_loans__loan_id__put","parameters":[{"required":true,"schema":{"title":"Loan Id","type":"integer"},"name":"loan_id","in":"path"},{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"query"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoanSchemaBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoanSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/loans/{loan_id}/month/{month}":{"get":{"tags":["Loans"],"summary":"Get Loan Summary","description":"Route to return the loan summary for a given month\nThe loan must be owned by or shared with the user","operationId":"get_loan_summary_loans__loan_id__month__month__get","parameters":[{"required":true,"schema":{"title":"Loan Id","type":"integer"},"name":"loan_id","in":"path"},{"required":true,"schema":{"title":"Month","type":"integer"},"name":"month","in":"path"},{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoanSummarySchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/loans/{loan_id}/share":{"post":{"tags":["Loans"],"summary":"Share Loan","description":"Route to share a loan with a user at user_id\nThe loan must be owned by the user at owner_id","operationId":"share_loan_loans__loan_id__share_post","parameters":[{"required":true,"schema":{"title":"Loan Id","type":"integer"},"name":"loan_id","in":"path"},{"required":true,"schema":{"title":"Owner Id","type":"integer"},"name":"owner_id","in":"query"},{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"LoanScheduleSchema":{"title":"LoanScheduleSchema","required":["month","open_balance","total_payment","principal_payment","interest_payment","close_balance"],"type":"object","properties":{"month":{"title":"Month","type":"integer"},"open_balance":{"title":"Open Balance","type":"number"},"total_payment":{"title":"Total Payment","type":"number"},"principal_payment":{"title":"Principal Payment","type":"number"},"interest_payment":{"title":"Interest Payment","type":"number"},"close_balance":{"title":"Close Balance","type":"number"}}},"LoanSchema":{"title":"LoanSchema","required":["amount","apr","term","owner_id","id"],"type":"object","properties":{"amount":{"title":"Amount","type":"number"},"apr":{"title":"Apr","type":"number"},"term":{"title":"Term","type":"integer"},"status":{"title":"Status","type":"string","default":"active"},"owner_id":{"title":"Owner Id","type":"integer"},"id":{"title":"Id","type":"integer"}}},"LoanSchemaBase":{"title":"LoanSchemaBase","required":["amount","apr","term","owner_id"],"type":"object","properties":{"amount":{"title":"Amount","type":"number"},"apr":{"title":"Apr","type":"number"},"term":{"title":"Term","type":"integer"},"status":{"title":"Status","type":"string","default":"active"},"owner_id":{"title":"Owner Id","type":"integer"}}},"LoanSummarySchema":{"title":"LoanSummarySchema","required":["current_principal","aggregate_principal_paid","aggregate_interest_paid"],"type":"object","properties":{"current_principal":{"title":"Current Principal","type":"number"},"aggregate_principal_paid":{"title":"Aggregate Principal Paid","type":"number"},"aggregate_interest_paid":{"title":"Aggregate Interest Paid","type":"number"}}},"UserSchema":{"title":"UserSchema","required":["username","id"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"id":{"title":"Id","type":"integer"}}},"UserSchemaBase":{"title":"UserSchemaBase","required":["username"],"type":"object","properties":{"username":{"title":"Username","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}